refactor(bbox-tube-temporal): rename Python package to bbox_tube_temporal
