+++ experiments/temporal-models/bbox-tube-temporal/docs/specs/2026-04-15-temporal-model-protocol-design.md
@@ -104,7 +104,7 @@ Batching summary:
 
 ## Components
 
-New files under `src/smokeynet_adapted/`:
+New files under `src/bbox_tube_temporal/`:
 
 - `model.py` - `SmokeynetTemporalModel(TemporalModel)` with
   `from_package()` factory and `predict()` method.
