+++ experiments/temporal-models/bbox-tube-temporal/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/smokeynet_adapted"]
+packages = ["src/bbox_tube_temporal"]
 
 [project]
 name = "bbox-tube-temporal"
