+++ experiments/temporal-models/bbox-tube-temporal/src/bbox_tube_temporal/model_input.py
