+++ experiments/temporal-models/bbox-tube-temporal/src/bbox_tube_temporal/types.py
