+++ experiments/temporal-models/bbox-tube-temporal/tests/test_lit_temporal.py
@@ -2,7 +2,7 @@
 
 import torch
 
-from smokeynet_adapted.lit_temporal import LitTemporalClassifier
+from bbox_tube_temporal.lit_temporal import LitTemporalClassifier
 
 
 def _batch(b: int = 2, t: int = 5) -> dict:
