+++ experiments/temporal-models/bbox-tube-temporal/tests/test_reproducibility.py
@@ -22,8 +22,8 @@ import torch
 from PIL import Image
 from torch.utils.data import DataLoader
 
-from smokeynet_adapted.dataset import TubePatchDataset
-from smokeynet_adapted.lit_temporal import LitTemporalClassifier
+from bbox_tube_temporal.dataset import TubePatchDataset
+from bbox_tube_temporal.lit_temporal import LitTemporalClassifier
 
 SEED = 1234
 OTHER_SEED = 5678
